mental_wellness_chatbot/requirements.txt
Flask==2.3.3
Flask-PyMongo==2.3.0
pymongo==4.6.1
google-generativeai==0.3.2
python-dotenv==1.0.0
Werkzeug==2.3.7
bcrypt==4.1.1
cohere